+++ packages/ragbits-document-search/tests/unit/test_web_source.py
@@ -1,17 +1,17 @@
 from sympy.testing import pytest
 
 from ragbits.document_search.documents.exceptions import SourceNotFoundError
-from ragbits.document_search.documents.sources.http import HttpSource
+from ragbits.document_search.documents.sources.web import WebSource
 
 
 def test_id():
-    source = HttpSource(url="http://example.com/file.pdf")
-    expected_id = "https:http://example.com/file.pdf"
+    source = WebSource(url="http://example.com/file.pdf")
+    expected_id = "web:http://example.com/file.pdf"
     assert source.id == expected_id
 
 
 async def test_from_uri_one_file():
-    file_uris = [
+    file_urls = [
         "http://www.without-secure-protocol.com/file.pdf",
         "https://www.with-secure-protocol.com/file.pdf",
         "https://www.longer-path.com/longer/path/to/file.pdf",
@@ -21,19 +21,19 @@ async def test_from_uri_one_file():
         "http://www.domain-only.com",
     ]
 
-    for uri in file_uris:
-        result = await HttpSource.from_uri(uri)
-        assert result[0] == HttpSource(url=uri)
+    for url in file_urls:
+        result = await WebSource.from_uri(url)
+        assert result[0] == WebSource(url=url)
 
 
 async def test_invalid_url_raises_exception():
-    wrong_uris = [
+    wrong_urls = [
         "some string",
         "https://www.url with spaces.com/path/to/file",
         "www.without-protocol.com/path/to/file",
         "http://www.domain-only-with-slash.com/",
     ]
 
-    for uri in wrong_uris:
+    for url in wrong_urls:
         with pytest.raises(SourceNotFoundError):
-            await HttpSource(url=uri).fetch()
+            await WebSource(url=url).fetch()