+++ packages/ragbits-conversations/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.10.1 (2025-03-19)
+
+### Changed
+
+- ragbits-core updated to version v0.10.1
+
 ## 0.10.0 (2025-03-17)
 
 ### Changed